other: merge all api into userAPIs file
